requirements.txt
django-haystack==3.0b1
whoosh
googlemaps
django-paypal
django-star-ratings
django-environ
